iGet/Set ConversationId for Streaming
fix

update
