+++ pkg/client/types.go
@@ -6,11 +6,46 @@ package symbl
 import (
 	"fmt"
 
+	rtinterfaces "github.com/dvonthenen/symbl-go-sdk/pkg/api/streaming/v1/interfaces"
+	cfginterfaces "github.com/dvonthenen/symbl-go-sdk/pkg/client/interfaces"
+	interfaces "github.com/dvonthenen/symbl-go-sdk/pkg/client/interfaces"
 	rest "github.com/dvonthenen/symbl-go-sdk/pkg/client/rest"
+	stream "github.com/dvonthenen/symbl-go-sdk/pkg/client/stream"
 )
 
 /*
-	Symbl REST API
+	REST Client
+*/
+type RestClient struct {
+	*rest.Client
+
+	creds *interfaces.Credentials
+	auth  *interfaces.AuthResp
+}
+
+/*
+	Streaming Client
+*/
+type StreamingOptions struct {
+	UUID           string
+	ProxyAddress   string
+	SymblConfig    *cfginterfaces.StreamingConfig
+	Callback       rtinterfaces.InsightCallback
+	SkipServerAuth bool
+}
+
+type StreamClient struct {
+	*stream.WebSocketClient
+
+	uuid           string
+	restClient     *RestClient
+	symblStreaming stream.WebSocketMessageCallback
+
+	options *StreamingOptions
+}
+
+/*
+	Symbl REST API Internals
 */
 type HeadersContext struct{}
